live-preview: Do not fail to render when top level sources are not loaded
Ths can open one file in VSCode and show the preview. Then you click
on another file in the file tree, which will replace the current file.

This replacement invalidates the data stored in the live-previews caches, so
it used to render an empty file. Read the data from disc instead -- if
that is an option.