Avoid inlining @image-url() expressions that can be expensive due to cache misses
If the `image-url()` expression of an `Image { source: @image-url("large-image.png"); ... }` gets inlined into geometry getters and other places that are called for every frame, then we might end up decoding images every frame, if the image isn't in the 5MB image decoder cache. It's better to rely on the `property <image>` of the `Image` for caching the decoded image, so don't inline those.

This fixes CPU being time being spent constantly on decoding images on the home automation lock screen.