Fix French translation not working in WASM build of printer demo
(For me this was also reproducible on desktop, but oddly not for
Olivier)

Letters like é were stripped from the noto sans font used by the demo,
and with FemtoVG we assume that presense of a character in the common
script implies full coverage (to avoid having to check every character).

Fix this by fixing the font itself. We wouldn't want those letters to
come from DejaVu :)