core: Improve forward/backward cursor movement in TextInput (#6990)
When a TextInput has a selection and a forward or backward movement
is performed (i.e. pressing Left or Right) without keeping the anchor
(i.e. Shift is not pressed), the selection collapses either to the
beginning or end of the selection based on the movement direction.

ChangeLog: Improved deselection behavior when pressing Left/Right in TextInput

Fixes #6511