live-preview: Move library closer to the 1.9 Reference docs
I can not exactly mirror the Reference manual as the library
is hard-coded to one level of depth, while the docs use two
levels.

This patch takes the "lower" level of the docs ("Basic Layouts", ...),
but groups all widgets below "Std Widgets" (the "upper" level of the docs
layout). This is because we see many projects that avoid our widgets,
so I wanted those users to be able to collapse all widgets with one click.

One additional refinement in this patch is that all widgets will always be
sorted after all builtin categories and that none of the builtin Windows
will get shown in the library as those are not really editable in the live
preview anyway.