Zephyr: improve render performance when using the NXP ELCDIF driver
When this driver is using the DMA API, we cannot do partial updates.
Previously, this code would then perform a full-screen DMA transfer for
each region the renderer changes. Now we make modifications to the pixel
data of the dirty regions, if required, and then do one fill-scren DMA
transfer of the final render buffer.