printerdemo_mcu PrinterQueueItem uses a status string
Use a string instead of the JobStatus enum.

This makes the PrinterQueue code compatible with the printerdemo code,
allowing us to switch between mcu and non-mcu versions of this demo,
depending on what will fit on a device.

A string is used, rather than porting the non-mcu demo to an enum,
because the non-mcu demo needs to work with the cpp_interpreted demo.