Make the Bazel build more robust
Ashpd doesn't compile out of the box because it uses zvariant-derive's
macro, which uses proc-macro-crate to find out if the zvariant types are
in `zvariant::` directly or indirectly via `zbus::zvariant::`. This is
checked by using proc-macro-crate to parse ashpd's Cargo.toml to see if
there's a dependency on zbus or zvariant, and if this fails the fallback
of zvariant-derive is to just use `zvariant::`. That's why before this
patch the Bazel build was injecting the zvariant dependency.

Instead of injecting this dependency to ashpd, it's better to fix
proc-macro-crate. This is less error prone in terms of version encoding
and it feels like the right approach. So the applied patch is also
submitted upstream as a PR.