Make the printer demo dark mode button easier to hit
Previously, the dark mode button was especially hard to hit on
embedded devices.

The TouchArea of the SideBarIcon used for this button was much smaller
than the others. By exposing a new property on the SideBarIcon:
icon-colorize, we can customize the colorize behaviour of the Image,
as well as the source. This allows us to re-use the internal Image for
the dark mode button. This allows us to use the default width, height
and x position of the icon. The touch areas of the sidebar buttons are
now uniform.