Docs: Fix <title> tag having docs version
The version was being set not just as a visual title, but also as part of the meta tag for every page on the docs site. Which will then change for each release as the version number changes.

Other sites have a static label here that is indepedent of the version e.g. Slint UI or Slint docs. 'Slint Docs' is now used.