Implement Sync for SharedVector and Weak
Note that this also make sure that a SharedVector for something that
isn't Sync isn't send. Otherwise we could send a `SharedVector<Cell<i32>>`
between threads, and because it is shared, different thread could modify
the same Cell at the same time, which is unsound.

Since SharedString is a wrapper to ShareVector, this will also add the
Sync impl there

ChangeLog: Fixed Sync and Send bound on SharedVector, SharedString, and Weak