Templates: Make a template for the blank issue
New GitHub interface add the Blank issue to the list of template
instead of being hidden in a small link
https://github.com/orgs/community/discussions/148713#discussioncomment-11868785
Following this change, we got several bad quality issues that didn't
follow the template.

Disable blank issue, but add a template for the blank issue with a title
that dissuade its use.