Fix panic when rendering with Skia and Metal for a long time
If CAMetalLayer decides to do double buffering for a long time, we'll keep increasing the age of the third buffer that we may have seen. We store that as u8, and this can overflow. Handle that case gracefully.

(This was introduced with commit d9404d4063eed0d60d8c3ca9b28b1a33f3ba2fd8)