Fix clippy warning
```
warning: passing a unit value to a function
   --> internal/interpreter/eval.rs:752:29
    |
752 | / ...                   item_rc_
753 | | ...                       .downcast::<corelib::items::ContextMenu>()
754 | | ...                       .unwrap()
755 | | ...                       .activated
756 | | ...                       .call(&(args[0].clone().try_into().unwrap(),))
757 | | ...                       .into()
    | |_________________________________^
    |
```

Activate, contrary to the other callback, doesn't return a value.
So write the Value::Void directly instead of using the fact that
`().into()` a value would produce Value::Void.

The code is more fragile as it will stop working if `activated` changed
return type and it no longer look consistant with the other call, but
let's make clippy happy anyway