Fix Gallery example sidebar accessibility (#7559)
The accessible-delegate-focus property is currently confused by the presence of the label at the top, since it expects an index in the accessible children list. Therefore the focused item would be off by one as far as assistive technologies are concerned.

I also had to move the FocusScope because accessible-delegate-focus is searched for in the ancestors of the element which is focused.

The tablist is now properly reported. Unfortunately these changes raise a more confusing bug: focus changes inside the tablist (using the arrow keys) don't seem to trigger a rebuild of the tree. When a tab is selected it is properly reported though. Even stranger: the focus changes are triggered if the focus is on the sidebar, when the toplevel window gains focus. Tabbing out of the sidebar and back on triggers the bug again though.