Material SpinBox: Fix binding loop in some layout
Having the width depending on the height is asking for trouble because
usually it is the height that depends on the width, so if the SpinBox is
getting in a complex layout, it causes a binding loop.
Workaround the problem by using the computed height.

This should fix the gallery build with the material style
(which is currently making the nightly build fail)