Build all data the Figma plugin shows without the CSS codegen (#7664)
The previous plugin built on top of the CSS codegen feature in figma. It would convert the CSS
output to Slint properties. This breaks down as there are CSS features that don't exist in Slint and
it lead to unexpected outputs.

This update generates all the Slint code from the raw Figma 'sceneNode' data. 
Right now Rectangle and Text nodes are supported. Brush color, linear and radial gradients.

It now produces more reliable code, but is probably missing a few features the CSS mode gave for free.

This also tidies up the way the Slint snippet is generated as there is a simple node -> slint workflow. As opposed 
to node -> CSS -> Slint + node -> Slint to work around places the CSS output was unhelpful.