Node.js: Remove windows x86 packages
Skia doesn't compile on windows x86 in the CI due the long paths that
exceed the 255 limit. Somehow the workaround that works for x64 doesn't
work here. Since this is the only windows x86 binary we produce
(not for lsp, vsce, python, cpp), I choose the easier way out and
remove the package.