Fix VSCE build
The CI has been using the deprecated vsce package to test .vsix creation, while it was renamed to @vscode/vsce a while ago. Last night, the vsce publish action we're using switched to @vscode/vsce and now the vsce build is failing, because of a bug that was hidden:

In vsce 2.25 support for `files` in `package.json` was introduced, which was just `bin/slint-lsp-*` in our package. That would omit the extension entry point and many other files, causing the vsce package creation step to fail.

Fix this by fixing the CI to use the same vsce version as the action is using, and replace the files with the default and the ignore setting of files we definitely want to exclude from the package.

With this, the package now contains these files:

```
slint-1.10.0.vsix
├─ [Content_Types].xml
├─ extension.vsixmanifest
└─ extension/
   ├─ .gitignore [0.03 KB]
   ├─ LICENSE.txt [0.04 KB]
   ├─ biome.json [0.2 KB]
   ├─ esbuild.js [2.22 KB]
   ├─ extension-logo.png [6.76 KB]
   ├─ language-configuration.json [1.49 KB]
   ├─ package.json [7.93 KB]
   ├─ readme.md [3.85 KB]
   ├─ slint-file-icon.svg [0.83 KB]
   ├─ slint.injection.json [1.61 KB]
   ├─ slint.markdown-injection.json [1.37 KB]
   ├─ slint.tmLanguage.json [8.79 KB]
   ├─ telemetry.json [2.21 KB]
   ├─ bin/
   │  └─ slint-lsp [82.21 MB]
   ├─ out/
   │  ├─ .gitignore [0 KB]
   │  ├─ browser.js [321.26 KB]
   │  ├─ browser.js.map [406.49 KB]
   │  ├─ browserServerMain.js [18.27 MB]
   │  ├─ browserServerMain.js.map [255.8 KB]
   │  ├─ extension.js [457.78 KB]
   │  ├─ extension.js.map [617.27 KB]
   │  ├─ slint_lsp_wasm.js [84.62 KB]
   │  └─ slint_lsp_wasm_bg.wasm [13.56 MB]
   ├─ snippets/
   │  └─ slint.json [0.53 KB]
   └─ static/
      └─ walkthroughs/
         └─ welcome/ (4 files) [17.1 KB]
```