live-preview: "Fix" types with units
We have several types that all support a range of units.
Internally all those units match back to one "base unit"
that the type is measured in.

In the spirit of simplicity: Support only that base unit
for now.

We need a more elaborate JSON scheme to pass units around
with those numbers :-/