live-preview: Update the live-preview data model
... on reloads. That is needed to make the UI aware
of the new data.

I tried to be clever and not update widgets on model
changes, but that is not a problem anymore as all
the property values are no longer passed through
the model. So the only reason to update the model now
is because the code was loaded fresh and we *need*
to refresh the entire UI in that case.