figma: Fix typescript errors not showing in-line for VScode (#7946)
Technically the Figma plugin is two typescript projects. A backend and a frontend. While typechecking via pnpm type-check works correctly, you would see no inline errors in VScode for the backend project. This lead to surpirses latert on when the CI runs the type-check script.

This fixes the issue so the experience inside VScode and from the type-check script match up.