Better handling of unset colors and brush properties (#8224)
Brush and color property widgets no longer show the default value of 'transparent' when a value has not been set. Instead the common symbol for no set color from other tools (a red diagonal line on a white background) is used.

When clicked the color picker is then shown and is populated with a default color that is 100% opaque.