Use Zustand to handle state and simplify Main.tsx (#8303)
This adds the Zustand state library to simplify react state handling.

It allows all the listening to the Figma side of the plugin to be moved inside it. This inturn allows quite a bit of logic to also move over to the store.ts file and start to leave Main.tsx to only describe the UI.