(Figma-plugin) Editing a property should update the snippet in design view (#8354)
In Figma dev-mode once you select an item in Figma it shows the slint snippet. If you then change any properties on the selected item the snippet updates. This is because the dev-mode automatically detects those changes and requests an update.

In the design view the api is different. The 'selection changed' event is used to show the initial snippet but subsequent changes will not be detected untill the user selects another item and then reselects the changed one. This PR adds a listener to the currently selected item to react to the NodeChanged event and update the snippet.

This PR also tidies up some dead async code. There were try catch blocks on code that according to the figma api docs never throw exceptions. Plus a few awaits on non async functions.