FemtoVG: Fix occasional panic on shutdown with wgpu
When closing the wgpu_texture example on Linux with Wayland,
we first suspend() the renderer and then drop it. Suspend() will
do the right thing and invoke the RenderingState::ReanderingTeardown
notifier, but the subsequent drop will try to do the same thing again,
and then we don't have wgpu device/instance/etc. anymore and the
attempt of unwrapping the graphics API panics.