live-preview: Disentangle threads
Make sure all code in preview is actually running in
the UI thread. For WASM this is a NOOP, as we have just
one thread there anyway.

For the native side, we need to handle all lsp to preview
messages in the UI thread. This is a bit trieckier than it
sounds: We do not want to bring u the UI thread fully till
the UI needs to be shown.

So re reuse the machanism used to switch between WASM preview
and native preview in VSCode:

We listen for a `ShowPreview` message and if the event loop
is not up, we bring it up and send a `RequestState` to the
LSP. That will respond with all the data we need and repeat
its `ShowPreview` rerquest afterwards.