wasm prevew: Make sure the call to show preview work
The lsp_to_preview_message is being sent from the JS runtime and
this is not called from the event loop.
When the preview is starting the 50ms timer to refresh itself,
since this is not called from the event loop, there is nothing to
wake the event loop up to fire this timer until a mouse event or some
other event wakes the event loop.
To make sure this works, call lsp_to_preview_message from the event
loop to make sure that the timer is already started when the event
loop goes to sleep and only sleep for the duration of that timer