interpreter: Make the From<ModelRc> for Value implement `set_row_data`
When converting a ModelRc<T> to a Value, the resulting model did not
implement set_row_data and therefore would not allow changing the model
from the code.
By adding a `TryFrom` bound, we can make sure set_row_data is
implemented.

This is technically a breaking change to add this bound, but most type
that implement From for value also implement TryFrom<Value>

These conversion function were added in
20443ec0df11b1eea6d5951be818b283cf012231 for Slint 1.9