POC for letting VScode open the browser help for basic elements
This allows the user to CTRL/CMD click on a basic element like 'Rectangle' or 'Text' and have the browser be opened to the correct help page.

There are many other ways this type of functionality could be done and probably in ways any editor using the Slint LSP can take advantage of. VScode also has multiple types of help, including the on-hover inline help system. But these all require more effort to support, including generating the right markdown files or teaching the LSP about the help. While this simple tweak gives a bit of extra help access to VScode users right away.