syntax_tests: allow to "bless" tests, and don't use a regexp
A regexp was used at the begining because I thought we would want to
allow error to contains things that were not predictable or that would
often change. But this is not the case. It is better to actually test
for the full error message

And when changing error message, a way to bless the test makes it easier